+++ testdata/de/dm/intellij/liferay/language/jsp/LiferayTaglibClassNameReferenceContributorTest/de/dm/Outer.java
@@ -1,0 +1,8 @@
+package de.dm;
+
+public interface Outer {
+
+    interface Inner {
+
+    }
+}